requirements.txt
gradio==5.6.0
langchain-community==0.3.7
langchain_core==0.3.19
langchain-openai==0.2.9
mysqlclient==2.2.6
pymysql==1.1.1
python-dotenv==1.0.1
tidb_vector==0.0.14
